migrate molecule card to use new ligand cards
